working version splitted in modules
